Update to 3.10.1
