+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nvngx"
-version = "0.1.3"
+version = "0.3.0"
 edition = "2021"
 authors = ["Victor Polevoy <maintainer@vpolevoy.com>"]
 description = "NVIDIA NGX bindings."
@@ -12,7 +12,7 @@ repository = "https://github.com/iddm/nvngx-rs"
 documentation = "https://docs.rs/nvngx-rs"
 
 [build-dependencies]
-bindgen = "0.69"
+bindgen = "0.71"
 
 [dependencies]
 widestring = "1"
